drop dma for being vapor. make testcmake.sh only re-run make when there is a Makefile.

+++ testcmake.sh
@@ -1,16 +1,19 @@
 #!/bin/sh
 # Copyright 2017-2025 Adrian Johnston
 
+# Only rerun cmake manually as is customary. Build artifacts ARE retained.
+
 export POSIXLY_CORRECT=1
 
 set -o errexit
 
-# Build artifacts are not retained.
+if [ ! -f "bin/Makefile" ]; then
 rm -rf ./bin; mkdir ./bin && cd ./bin
-
-set -x
-
 cmake ..
+else
+cd ./bin
+fi
+
 make
 ./hxtest
